+++ fmn/rules/fedbadges.py
@@ -8,6 +8,16 @@ def fedbadges_badge_award(config, message):
     return message['topic'].endswith('fedbadges.badge.award')
 
 
+def fedbadges_person_first_login(config, message):
+    """ Badges: Someone logs in for the first time
+
+    Adding this rule will let through notifications from the `Fedora Badges
+    <https://badges.fedoraproject.org>`_ system whenever someone *logs in*
+    for the *first time*.
+    """
+    return message['topic'].endswith('fedbadges.person.login.first')
+
+
 def fedbadges_person_rank_advance(config, message):
     """ Badges: The rank of someone changed on the badges leaderboard